// src/app/app.interface.ts
export class Details {

    positionid: string;
    effectivedate: string;
    status: string;
    legalentity: string;
    bussinessunit: string;
    lobsegment: string;
    vertical: string;
    subvertival: string;
    costcenter: string;
    Geozone: string;
    facility: string;
    employeeclass: string;
    employeetype: string;
    jobfamily: string;
    Jobcode: string;
    internalspecilization: string;
    variablepaytype: string;
    reportingmanager: string;
    buhrspoc: string;

}